refactor: create 폴더 이름 변경 및 라우터 변경
