+++ frontend/src/pages/host/spaceCreate/funnel/SpaceCreateFunnel.styles.ts
