+++ frontend/src/pages/host/spaceCreate/funnel/funnel.validators.ts
