+++ frontend/src/pages/host/spaceCreate/funnel/funnelBasePage/FunnelBasePage.styles.ts
