+++ frontend/src/pages/host/spaceCreate/funnelElements/spaceCheckElement/SpaceCheckElement.styles.ts
