+++ frontend/src/pages/host/spaceCreate/funnelElements/spaceCheckElement/SpaceCheckElement.tsx
@@ -1,9 +1,9 @@
+import defaultImage from '../../../../../@assets/images/default-image.png';
 import InfoRow from '../../../../../components/host/infoRow/InfoRow';
 import type {
   CreateFunnelForm,
   FunnelElementProps,
 } from '../../../../../types/funnel.type';
-import defaultImage from '../../../../@assets/images/default-image.png';
 import { Thumbnail } from '../../../../MainPage.common.styles';
 import FunnelBasePage from '../../funnel/funnelBasePage/FunnelBasePage';
 import * as S from './SpaceCheckElement.styles';
