+++ frontend/src/pages/host/spaceCreate/funnelElements/spaceDetailElement/SpaceDetailElement.styles.tsx
