+++ frontend/src/pages/host/spaceCreate/funnelElements/spaceDetailElement/SpaceDetailElement.tsx
