+++ frontend/src/router/router.tsx
@@ -1,9 +1,9 @@
 import { createBrowserRouter } from 'react-router-dom';
 import Layout from '../components/layout/global/layout/Layout';
-import SpaceCreateFunnel from '../pages/create/funnel/SpaceCreateFunnel';
 import GuestMainPage from '../pages/guest/mainpage/GuestMainPage';
 import Dashboard from '../pages/host/dashboard/Dashboard';
 import HostMainPage from '../pages/host/mainPage/HostMainPage';
+import SpaceCreateFunnel from '../pages/host/spaceCreate/funnel/SpaceCreateFunnel';
 import SpaceEditPage from '../pages/host/spaceEditPage/SpaceEditPage';
 import MainPage from '../pages/MainPage';
 import type { AppRouteObject } from '../types/route.type';
@@ -41,6 +41,13 @@ const routes: AppRouteObject[] = [
               headerIcons: ['settings'],
             },
           },
+          {
+            path: 'create-space',
+            element: <SpaceCreateFunnel />,
+            handle: {
+              noHeader: true,
+            },
+          },
         ],
       },
       {
@@ -51,13 +58,6 @@ const routes: AppRouteObject[] = [
             element: <GuestMainPage />,
           },
         ],
-      },
-      {
-        path: 'create',
-        element: <SpaceCreateFunnel />,
-        handle: {
-          noHeader: true,
-        },
       },
     ],
   },
